backend_v2/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.34.0
numpy>=2.1.0
pandas>=2.2.0
vnstock>=3.5.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
aiomysql>=0.2.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
tenacity>=9.0.0
httpx>=0.24.0
redis>=5.0.0
pyarrow>=15.0.0
